feat(tree): convert to TypeScript and export types
